Db/Spell : Desperate isnt Scaling With Spell Power
The Skill http://www.wowhead.c...rayer#changelog of priest is bugged.
Inscription- Instantly heal the caster for 30% of their total health
Bug- The skill is healing the 30% it should, plus the spell power of the priest.
Example: I have a priest with 152029 Hp (30% of 152029 is 45609, so, it should heal 45609, but with the bug of the spell power, is healing 53785)